requirements.txt
streamlit
streamlit-chat
openai
azure-identity
python-docx
python-dotenv
pydantic

